+++ src/Media/Media.js
@@ -1,10 +1,37 @@
-// @flow
-import elem from '../elem-factory'
-import withTheme from '../withTheme'
+import React, { createContext, useState, useEffect } from 'react'
 
-const Media = withTheme('Media', props => elem({ classes: 'media', ...props }))
+/**
+ * Media Context
+ */
+export const Context = createContext()
 
-const Body = withTheme('MediaBody', props => elem({ classes: 'media-body', ...props }))
-Media.Body = Body
+/**
+ * The Media provider accepts breakpoints and can be used with the `useMedia`
+ * hook in any component Default breakpoints are:
+ *
+ * - sm = >= 0
+ * - md = >= 900
+ * - lg = >= 1250
+ */
+const MediaProvider = ({ children, breakpoints = [0, 900, 1250] }) => {
+  const calcBreakpoints = w => ({
+    sm: w < breakpoints[1],
+    md: w >= breakpoints[1] && w < breakpoints[2],
+    lg: w >= breakpoints[2],
+  })
 
-export default Media
+  const [bps, updateBps] = useState(calcBreakpoints(window.innerWidth))
+  const setBreakpoints = () => {
+    updateBps(calcBreakpoints(window.innerWidth))
+  }
+
+  useEffect(() => {
+    breakpoints.forEach(bp => {
+      const mq = window.matchMedia(`(min-width: ${bp}px)`)
+      mq.addListener(setBreakpoints)
+    })
+  }, [])
+
+  return <Context.Provider value={bps}> {children}</Context.Provider>
+}
+export default MediaProvider